feat: SQLite driver should emit "connection" event when new connection is created
